controller: Support learning routes per iface.
Previously we just assumed that if a LR had multiple LRPs bound to the
local chassis that all routes where valid for all of these LRPs.
This commit handles a previous option in ovn-nb that allowes the user to
specify the interface name that a route needs to use to be acceptable
for a given LRP.
The users can then have a 1:1 relationship between LRPs and interfaces
to allow ovn-controller to determine which route belongs to which LRP.

This can e.g. be used if a chassis has multiple uplinks. Each of these
uplinks could be assigned to a different LRP of the same LR connected to
the same localnet.
The option allows then to map e.g. a learned default route to each
individual LRP. If one of the links would go down just one of the
learned routes would be removed, causing ECMP across multiple chassis to
be correctly weighted.

Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>