controller: Support learning routes per iface.
Previously we just assumed that if a LR had multiple LRPs bound to the
local chassis that all routes where valid for all of these LRPs.
This commit handles a previous option in ovn-nb that allowes the user to
specify the interface name that a route needs to use to be acceptable
for a given LRP.
The user specifies a port name in the northbound database. If this port
is bound locally we use the interface name associated with it for
filtering routes.
Additionally users can overwrite that interface name to also support
cases where there is no referencable port name in the northbound.

This can e.g. be used if a chassis has multiple uplinks. Each of these
uplinks could be assigned to a different LRP of the same LR connected to
the same localnet.
The option allows then to map e.g. a learned default route to each
individual LRP. If one of the links would go down just one of the
learned routes would be removed, causing ECMP across multiple chassis to
be correctly weighted.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>