northd: Allow LB health checks from router IPs.
Load Balancer Health Checks require specifying "source IP". This IP
has to be from the subnet of the monitored LB backend and should not
be used by any other port in the LSP. If the "source IP" is also used
by some other port, the host with the LB backend won't be able to communicate
with this port due to the ARP responder rule installed by the controller.

This limitation forces CMSes to reserve an IP per LSP with LB backend, which
is not always practical or possible.

This change allows usage of LRP IPs as the source of LB health check
probes. It introduces following changes for such scenario:

  * service_monitor (SBDB) will use MAC address of the LRP
  * ARP responder rule is not necessary in this case
  * Destination IP, "inport" and source port will be used to determine that a
    packet is a response to a health check probe. These packets will be
    redirected to the controller.

Behavior is unchanged for the LB health checks that use reserved/unused
"source IPs".

Signed-off-by: Martin Kalcok <martin.kalcok@canonical.com>